requirements.txt
alabaster==0.7.12
astroid==2.4.2
attr==0.3.1
attrs==20.3.0
Babel==2.9.0
beautifulsoup4==4.9.3
blis==0.7.4
catalogue==2.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coremltools==4.1
cymem==2.0.5
docutils==0.16
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.0.0/fr_core_news_sm-3.0.0-py3-none-any.whl
future==0.18.2
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
isort==5.7.0
Jinja2==2.11.3
joblib==1.0.0
jsonschema==3.2.0
kraken==2.0.8
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
mpmath==1.2.1
murmurhash==1.0.5
numpy==1.19.5
packaging==20.9
pathy==0.3.6
Pillow==8.2.0
pkg-resources==0.0.0
pluggy==0.13.1
preshed==3.0.5
protobuf==3.15.8
py==1.10.0
pydantic==1.7.3
Pygments==2.8.0
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
python-bidi==0.4.2
python-Levenshtein==0.12.2
pytz==2021.1
regex==2021.4.4
requests==2.25.1
scikit-learn==0.19.2
scipy==1.6.0
six==1.15.0
smart-open==3.0.0
snowballstemmer==2.1.0
soupsieve==2.1
spacy==3.0.3
spacy-legacy==3.0.1
Sphinx==3.5.1
sphinx-rtd-theme==0.5.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
srsly==2.4.0
sympy==1.8
termcolor==1.1.0
thinc==8.0.1
threadpoolctl==2.1.0
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tqdm==4.56.2
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wasabi==0.8.2
wrapt==1.12.1
